refactor(test): separate tests
