+++ src/__test__/prepare.js
@@ -1,0 +1,3 @@
+process.on('unhandledRejection', (reason) => {
+  throw reason;
+});
